Rename xml view test to scheme
