+++ test/e2e/src/pages/bpmnXmlView.ts
